Remove duplicate settlement example.
