+++ contracts/settlements/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "settlements"
 version = "3.0.0"
-authors = ["Polymesh Association"]
+authors = ["PolymeshAssociation"]
 edition = "2021"
 publish = false
 
@@ -11,7 +11,7 @@ ink = { version = "4.3", default-features = false }
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2", default-features = false, features = ["derive"], optional = true }
 
-polymesh-api = { version = "3.5", default-features = false, features = ["ink"] }
+polymesh-ink = { version = "3.0.0", path = "../upgradeable-polymesh-ink/", default-features = false, features = ["as-library"] }
 
 [lib]
 path = "lib.rs"
@@ -22,6 +22,6 @@ std = [
     "ink/std",
     "scale/std",
     "scale-info/std",
-		"polymesh-api/std",
+    "polymesh-ink/std",
 ]
 ink-as-dependency = []
