--- contracts/upgradeable-polymesh-ink/examples/settlements/.gitignore
@@ -1,5 +1,0 @@
-# Ignore build artifacts from the local tests sub-crate.
-/target/
-
-# Ignore backup files creates by cargo fmt.
-**/*.rs.bk
