--- contracts/upgradeable-polymesh-ink/examples/settlements/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "settlements"
-version = "3.0.0"
-authors = ["PolymeshAssociation"]
-edition = "2021"
-publish = false
-
-[dependencies]
-ink = { version = "4.3", default-features = false }
-
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2", default-features = false, features = ["derive"], optional = true }
-
-polymesh-ink = { version = "3.0.0", path = "../../", default-features = false, features = ["as-library"] }
-
-[lib]
-path = "lib.rs"
-
-[features]
-default = ["std"]
-std = [
-    "ink/std",
-    "scale/std",
-    "scale-info/std",
-    "polymesh-ink/std",
-]
-ink-as-dependency = []
